+++ src/nemo-image-properties-page.c
@@ -480,7 +480,7 @@ file_read_callback (GObject      *object,
 
 	if (count_read > 0) {
 
-		g_assert (count_read <= sizeof(page->details->buffer));
+		g_assert (count_read <= (gssize) sizeof(page->details->buffer));
 
 #ifdef HAVE_EXIF
 		exif_still_loading = exif_loader_write (page->details->exifldr,
